requirements.txt
celery
dj-database-url==0.4.0
djangorestframework
django-filters
Django==1.9.2
gunicorn==19.4.5
mock
psycopg2==2.6.1
requests
selenium
whitenoise==2.0.6
